+++ web/frontend-v2/openapi-ts.config.ts
@@ -1,14 +1,6 @@
 import { defineConfig } from '@hey-api/openapi-ts';
 
 export default defineConfig({
-  client: false,
   input: 'openapi.json',
-  output: {
-    path: 'src/generated',
-    format: 'prettier',
-  },
-  types: {
-    enums: 'javascript',
-  },
-  services: false,
+  output: 'src/generated',
 });